+++ app-shell/src/discovery.ts
@@ -20,7 +20,6 @@ import {
   CLEAR_CACHE,
 } from '@opentrons/app/src/redux/discovery/actions'
 import { OPENTRONS_USB } from '@opentrons/app/src/redux/discovery/constants'
-import { handleNotificationConnectionsFor } from './notify'
 
 import { getFullConfig, handleConfigChange } from './config'
 import { createLogger } from './log'
@@ -35,6 +34,7 @@ import type {
 
 import type { Action, Dispatch } from './types'
 import type { ConfigV1 } from '@opentrons/app/src/redux/config/schema-types'
+import { handleNotificationConnectionsFor } from './notifications'
 
 const log = createLogger('discovery')
