+++ app-shell/src/notifications/log.ts
@@ -1,0 +1,3 @@
+import { createLogger } from '../log'
+
+export const notifyLog = createLogger('notify')
